% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_process_subfuncs.R
\name{compute_event_counts}
\alias{compute_event_counts}
\title{Compute the counts of each event within the patient record and across the cohort}
\usage{
compute_event_counts(
  cohort,
  grouped_list,
  site_col,
  time = FALSE,
  event_tbl = read_codeset("pes_events_1", "ccccc")
)
}
\arguments{
\item{cohort}{table of cohort members with at least the site_col, \code{person_id}, \code{start_date}, and \code{end_date}}

\item{grouped_list}{list of columns that should be used to group the table}

\item{site_col}{the name of the column with the site names}

\item{time}{logical to determine whether to output the check across time}

\item{event_tbl}{a table with the definitions for each event with the columns

\if{html}{\out{<div class="sourceCode">}}\preformatted{             - @event - A or B
             - @event_label - a descriptive label for the event
             - @domain_tbl - the default CDM table from which data is retrieved
             - @concept_field - the field in the table where the codes of interest are
                                stored
             - @date_field - the date field to be used to establish the index &
                             occurrence dates
             - @codeset_name - the name of the codeset in the specs directory to
                               define the variable of interest
             - @filter_logic - a string indicating any filter logic that should be
                               applied to establish the event

                               ex: an Hba1c > 6.5
}\if{html}{\out{</div>}}}
}
\value{
a dataframe with the proportion of patients that have event A, have event B,
have both events, or have neither event within their patient record
}
\description{
Compute the counts of each event within the patient record and across the cohort
}
