% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_process_subfuncs_pcornet.R
\name{compute_event_counts_pcnt}
\alias{compute_event_counts_pcnt}
\title{Compute the counts of each event within the patient record and across the cohort}
\usage{
compute_event_counts_pcnt(
  cohort,
  grouped_list,
  site_col,
  time = FALSE,
  event_tbl = patientrecordconsistency::prc_event_file
)
}
\arguments{
\item{cohort}{table of cohort members with at least the site_col, \code{patid}, \code{start_date}, and \code{end_date}}

\item{grouped_list}{list of columns that should be used to group the table}

\item{site_col}{the name of the column with the site names}

\item{time}{logical to determine whether to output the check across time}

\item{event_tbl}{a table with the definitions for each event with the columns
\itemize{
\item \code{event} - A or B
\item \code{event_label} - a descriptive label for the event
\item \code{domain_tbl} - the default CDM table from which data is retrieved
\item \code{concept_field} - the field in the table where the codes of interest are stored
\item \code{date_field} - the date field to be used to establish the index & occurrence dates
\item \code{vocabulary_field} - (PCORnet only) The name of the column in the domain table where the vocabulary type is stored
\item \code{codeset_name} - the name of the codeset in the specs directory to define the variable of interest
\item \code{filter_logic} - a string indicating any filter logic that should be applied to establish the event
ex: an Hba1c > 6.5
}}
}
\value{
a dataframe with the proportion of patients that have event A, have event B,
have both events, or have neither event within their patient record
}
\description{
Compute the counts of each event within the patient record and across the cohort
}
