% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prc_event_file}
\alias{prc_event_file}
\title{PRC Sample Event File}
\format{
\subsection{prc_event_file}{

A data frame with 2 rows and 7 columns
\describe{
\item{event}{A or B; represents which of the two events should occur first and second in the sequence}
\item{event_label}{An arbitrary string to label what the event is}
\item{domain_tbl}{The name of the CDM table associated with the domain where concepts associated with the event can be found}
\item{concept_field}{The name of the column in the domain table that contains the concepts of interest}
\item{date_field}{The name of the column in the domain table that contains dates to be used for establishing the sequence of events.}
\item{vocabulary_field}{(PCORnet only) The name of the column in the domain table where the vocabulary type is stored}
\item{codeset_name}{The name of the codeset (found in the file_subdirectory) to be used to define the event}
\item{filter_logic}{(optional) a string to be parsed as logic to filter the domain_tbl as needed to best represent the event}
}
}
}
\usage{
prc_event_file
}
\description{
A sample version of the file structure expected for the \code{prc_event_file}
parameter in the \code{prc_process} function. The user should recreate this file
and include their own domain definitions.
}
\details{
Files referenced in the \code{codeset_name} column should be kept in the \code{file_subdirectory}
identified when \code{initialize_dq_session} is called.
}
\keyword{datasets}
