% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_output_subfuncs.R
\name{prc_ms_anom_nt}
\alias{prc_ms_anom_nt}
\title{\emph{Multi Site, Anomaly Detection, No Time}}
\usage{
prc_ms_anom_nt(process_output)
}
\arguments{
\item{process_output}{the output from \code{prc_process}}
}
\value{
a dot plot where the shape of the dot represents whether the point is
anomalous, the color of the dot represents the jaccard similarity index
for a given follow up window, and the size of the dot represents the mean index
across all sites

SD: ', round(sd_val,2),
SD: ', round(sd_val,2),
}
\description{
\emph{Multi Site, Anomaly Detection, No Time}
}
