% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_output.R
\name{prc_output}
\alias{prc_output}
\title{Patient Record Consistency Visualizations}
\usage{
prc_output(
  process_output,
  output_function,
  dist_from_stat = "mean",
  event_filter
)
}
\arguments{
\item{process_output}{the summary output from \code{prc_process}}

\item{output_function}{the name of the output function that should be executed; this is provided
in a console message after \code{prc_process} finishes executing}

\item{dist_from_stat}{the statistic from which distance should be measured for ms_exp_at
acceptable values are \code{mean} or \code{median}}

\item{event_filter}{the event type of interest for the analysis for ss_anom_at or ms_anom_at;
can be either \code{a}, \code{b}, \code{both}, or \code{neither}}
}
\value{
a graph visualizing the output from \code{prc_process}; see individual functions for
more specific details
}
\description{
Patient Record Consistency Visualizations
}
