% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_process_omop.R
\name{prc_process_omop}
\alias{prc_process_omop}
\title{Patient Record Consistency -- OMOP}
\usage{
prc_process_omop(
  cohort,
  prc_event_file,
  multi_or_single_site = "single",
  anomaly_or_exploratory = "exploratory",
  age_groups = NULL,
  patient_level_tbl = FALSE,
  fu_breaks = c(0, 1, 3, 8, 11, 15, 25, 50, 100),
  p_value = 0.9,
  time = FALSE,
  time_span = c("2012-01-01", "2020-01-01"),
  time_period = "year"
)
}
\arguments{
\item{cohort}{cohort for SQUBA testing; required fields:
\itemize{
\item \code{site}
\item \code{person_id}
\item \code{start_date}
\item \code{end_date}
}}

\item{prc_event_file}{a table with the definitions for each event with the columns
\itemize{
\item \code{event} - A or B
\item \code{event_label} - a descriptive label for the event
\item \code{domain_tbl} - the default CDM table from which data is retrieved
\item \code{concept_field} - the field in the table where the codes of interest are stored
\item \code{date_field} - the date field to be used to establish the index & occurrence dates
\item \code{vocabulary_field} - (PCORnet only) The name of the column in the domain table where the vocabulary type is stored
\item \code{codeset_name} - the name of the codeset in the specs directory to define the variable of interest
\item \code{filter_logic} - a string indicating any filter logic that should be applied to establish the event
ex: an Hba1c > 6.5
}}

\item{multi_or_single_site}{Option to run the function on a single vs multiple sites
\itemize{
\item \code{single} - run the function for a single site
\item \code{multi} - run the function for multiple sites
}}

\item{anomaly_or_exploratory}{Option to conduct an exploratory or anomaly detection analysis. Exploratory analyses give a high
level summary of the data to examine the fact representation within the cohort. Anomaly detection
analyses are specialized to identify outliers within the cohort.}

\item{age_groups}{If you would like to stratify the results by age group,  create a table or CSV file with the following
columns and include it as the \code{age_groups} function parameter:
\itemize{
\item \code{min_age}: the minimum age for the group (i.e. 10)
\item \code{max_age}: the maximum age for the group (i.e. 20)
\item \code{group}: a string label for the group (i.e. 10-20, Young Adult, etc.)
}

If you would \emph{not} like to stratify by age group, leave the argument as NULL}

\item{patient_level_tbl}{logical controlling whether patient level output is returned or not}

\item{fu_breaks}{a numeric vector that defines how to group different windows of follow up time}

\item{p_value}{the p value to be used as a threshold in the multi-site anomaly detection analysis}

\item{time}{a logical that tells the function whether you would like to look at the output over time}

\item{time_span}{when time = TRUE, this argument defines the start and end dates for the time period of interest. should be
formatted as c(start date, end date) in yyyy-mm-dd date format}

\item{time_period}{when time = TRUE, this argument defines the distance between dates within the specified time period. defaults
to \code{year}, but other time periods such as \code{month} or \code{week} are also acceptable}
}
\value{
a dataframe summarizing how often two events occur and co-occur within a patient record
}
\description{
Patient Record Consistency -- OMOP
}
\keyword{internal}
