% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_output_subfuncs.R
\name{prc_ss_anom_cs}
\alias{prc_ss_anom_cs}
\title{\emph{Single Site, Anomaly Detection, Cross-Sectional}}
\usage{
prc_ss_anom_cs(process_output)
}
\arguments{
\item{process_output}{the output from \code{prc_process}}
}
\value{
a bar graph displaying the jaccard similarity index for the two events
within each user defined follow up window
}
\description{
\emph{Single Site, Anomaly Detection, Cross-Sectional}
}
