% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_output_subfuncs.R
\name{prc_ss_anom_la}
\alias{prc_ss_anom_la}
\title{\emph{Single Site, Anomaly Detection, Longitudinal}}
\usage{
prc_ss_anom_la(process_output, event_filter, facet = NULL)
}
\arguments{
\item{process_output}{the output from \code{prc_process}}

\item{event_filter}{the event type of interest for the analysis; can be either
\code{a}, \code{b}, \code{both}, or \code{neither}}

\item{facet}{vector of variables to be used to facet the graph}
}
\value{
if analysis was executed by year or greater, a P Prime control chart
is returned with outliers marked with orange dots

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if analysis was executed by month or smaller, an STL regression is
    conducted and outliers are marked with red dots. the graphs representing
    the data removed in the regression are also returned
}\if{html}{\out{</div>}}
}
\description{
\emph{Single Site, Anomaly Detection, Longitudinal}
}
