% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_output_subfuncs.R
\name{prc_ss_exp_cs}
\alias{prc_ss_exp_cs}
\title{\emph{Single Site, Exploratory, Cross-Sectional}}
\usage{
prc_ss_exp_cs(process_output)
}
\arguments{
\item{process_output}{the output from \code{prc_process}}
}
\value{
a bar graph displaying the proportion (left y-axis) and raw count (right y-axis) of
patients who have only event a, only event b, both events, or neither event present
in their record
}
\description{
\emph{Single Site, Exploratory, Cross-Sectional}
}
