% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_output_subfuncs.R
\name{prc_ss_exp_la}
\alias{prc_ss_exp_la}
\title{\emph{Single Site, Exploratory, Longitudinal}}
\usage{
prc_ss_exp_la(process_output)
}
\arguments{
\item{process_output}{the output from \code{prc_process}}
}
\value{
a line graph the displaying the proportion of patients with only event a,
only event b, both events, or neither event across time
}
\description{
\emph{Single Site, Exploratory, Longitudinal}
}
