% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_output_subfuncs.R
\name{prc_ms_anom_la}
\alias{prc_ms_anom_la}
\title{\emph{Multi Site, Anomaly Detection, Longitudinal}}
\usage{
prc_ms_anom_la(
  process_output,
  event_filter,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{the output from \code{prc_process}}

\item{event_filter}{the event type of interest for the analysis; can be either
\code{a}, \code{b}, \code{both}, or \code{neither}}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}
}
\value{
three graphs:
\enumerate{
\item line graph that shows the smoothed proportion of patients
across time computation with the Euclidean distance associated with each line
\item line graph that shows the raw proportion of patients
across time computation with the Euclidean distance associated with each line
\item a bar graph with the Euclidean distance value for each site, with the average
proportion as the fill
}
}
\description{
\emph{Multi Site, Anomaly Detection, Longitudinal}
}
