% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_output_subfuncs.R
\name{prc_ms_anom_nt}
\alias{prc_ms_anom_nt}
\title{\emph{Multi Site, Anomaly Detection, No Time}}
\usage{
prc_ms_anom_nt(process_output)
}
\arguments{
\item{process_output}{the output from \code{prc_process}}
}
\value{
a dot plot where the shape of the dot represents whether the point is
anomalous, the color of the dot represents the jaccard similarity index
for a given follow up window, and the size of the dot represents the mean index
across all sites

\if{html}{\out{<div class="sourceCode">}}\preformatted{    if there were no groups eligible for analysis, a heat map showing the jaccard index
    and a dot plot showing each site's average standard deviation away from the mean
    index is returned instead
}\if{html}{\out{</div>}}

SD: ', round(sd_val,2),
SD: ', round(sd_val,2),
}
\description{
\emph{Multi Site, Anomaly Detection, No Time}
}
