% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_output_subfuncs.R
\name{prc_ms_exp_cs}
\alias{prc_ms_exp_cs}
\title{\emph{Multi Site, Exploratory, Cross-Sectional}}
\usage{
prc_ms_exp_cs(process_output, large_n = FALSE, large_n_sites = NULL)
}
\arguments{
\item{process_output}{the output from \code{prc_process}}

\item{large_n}{a boolean indicating whether the large N visualization, intended for a high
volume of sites, should be used; defaults to FALSE}

\item{large_n_sites}{a vector of site names that can optionally generate a filtered visualization}
}
\value{
a bar graph displaying the proportion of patients at each site with
only event a, only event b, both events, or neither event present in
their record
}
\description{
\emph{Multi Site, Exploratory, Cross-Sectional}
}
