% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_output.R
\name{prc_output}
\alias{prc_output}
\title{Patient Record Consistency -- Output Generation}
\usage{
prc_output(
  process_output,
  dist_from_stat = "mean",
  event_filter = NULL,
  large_n = FALSE,
  large_n_sites = NULL
)
}
\arguments{
\item{process_output}{\emph{tabular input} || \strong{required}

The tabular output produced by \code{pes_process}

Note any patient-level results generated are not intended to be used with this function.}

\item{dist_from_stat}{\emph{string} || defaults to \code{mean}

A string indicating the statistic from which distance should be measured for the
\verb{Multi-Site, Exploratory, Longitudinal} check

Acceptable values are \code{mean} or \code{median}}

\item{event_filter}{\emph{string} || defaults to \code{NULL}

A string indicating the event co-occurrence type of interest for the analysis.
This parameter is required for the following checks:
\itemize{
\item \verb{Single Site, Anomaly Detection, Longitudinal}
\item \verb{Multi-Site, Anomaly Detection, Longitudinal}
}

Acceptable values are \code{a}, \code{b}, \code{both}, or \code{neither}}

\item{large_n}{\emph{boolean} || defaults to \code{FALSE}

For Multi-Site analyses, a boolean indicating whether the large N
visualization, intended for a high volume of sites, should be used. This
visualization will produce high level summaries across all sites, with an
option to add specific site comparators via the \code{large_n_sites} parameter.}

\item{large_n_sites}{\emph{vector} || defaults to \code{NULL}

When \code{large_n = TRUE}, a vector of site names that can add site-level information
to the plot for comparison across the high level summary information}
}
\value{
This function will produce a graph to visualize the results
from \code{prc_process} based on the parameters provided. The default
output is typically a static ggplot or gt object, but interactive
elements can be activated by passing the plot through \code{make_interactive_squba}.
For a more detailed description of output specific to each check type,
see the PEDSpace metadata repository
}
\description{
Using the tabular output generated by \code{prc_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_squba()}
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'patientrecordconsistency'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'prc_process_test',
                      working_directory = my_directory,
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = my_file_folder,
                      cdm_schema = NA)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000), # RSQLite does not store date objects,
                                      # hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Build function input table
prc_events <- tidyr::tibble(event = c('a', 'b'),
                            event_label = c('hypertension', 'inpatient/ED visit'),
                            domain_tbl = c('condition_occurrence', 'visit_occurrence'),
                            concept_field = c('condition_concept_id', 'visit_concept_id'),
                            date_field = c('condition_start_date', 'visit_start_date'),
                            vocabulary_field = c(NA, NA),
                            codeset_name = c('dx_hypertension', 'visit_edip'),
                            filter_logic = c(NA, NA))

#' Execute `prc_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
prc_process_example <- prc_process(cohort = cohort,
                                   multi_or_single_site = 'single',
                                   anomaly_or_exploratory = 'exploratory',
                                   time = FALSE,
                                   omop_or_pcornet = 'omop',
                                   prc_event_file = prc_events) \%>\%
  suppressMessages()

prc_process_example

#' Execute `prc_output` function
prc_output_example <- prc_output(process_output = prc_process_example)

prc_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_squba()`

make_interactive_squba(prc_output_example)
}
