% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prc_output.R
\name{prc_output}
\alias{prc_output}
\title{Patient Record Consistency -- Output Generation}
\usage{
prc_output(
  process_output,
  output_function,
  dist_from_stat = "mean",
  event_filter = NULL
)
}
\arguments{
\item{process_output}{the summary output from \code{prc_process}}

\item{output_function}{the name of the output function that should be executed; this is provided
in a console message after \code{prc_process} finishes executing}

\item{dist_from_stat}{the statistic from which distance should be measured for ms_exp_at
acceptable values are \code{mean} or \code{median}}

\item{event_filter}{the event type of interest for the analysis for ss_anom_at or ms_anom_at;
can be either \code{a}, \code{b}, \code{both}, or \code{neither}}
}
\value{
a graph visualizing the output from \code{prc_process}; see individual functions for
more specific details
}
\description{
Using the tabular output generated by \code{prc_process}, this function will build a graph to
visualize the results. Each function configuration will output a bespoke ggplot. Theming can
be adjusted by the user after the graph has been output using \code{+ theme()}. Most graphs can
also be made interactive using \code{make_interactive_ssdqa()}
}
\examples{

#' Source setup file
source(system.file('setup.R', package = 'patientrecordconsistency'))

#' Create in-memory RSQLite database using data in extdata directory
conn <- mk_testdb_omop()

#' Establish connection to database and generate internal configurations
initialize_dq_session(session_name = 'prc_process_test',
                      working_directory = getwd(),
                      db_conn = conn,
                      is_json = FALSE,
                      file_subdirectory = system.file('extdata',
                                        package = 'patientrecordconsistency'),
                      cdm_schema = NA)

#' Build mock study cohort
cohort <- cdm_tbl('person') \%>\% dplyr::distinct(person_id) \%>\%
  dplyr::mutate(start_date = as.Date(-5000), # RSQLite does not store date objects,
                                      # hence the numerics
                end_date = as.Date(15000),
                site = ifelse(person_id \%in\% c(1:6), 'synth1', 'synth2'))

#' Build function input table
prc_events <- tidyr::tibble(event = c('a', 'b'),
                            event_label = c('hypertension', 'inpatient/ED visit'),
                            domain_tbl = c('condition_occurrence', 'visit_occurrence'),
                            concept_field = c('condition_concept_id', 'visit_concept_id'),
                            date_field = c('condition_start_date', 'visit_start_date'),
                            vocabulary_field = c(NA, NA),
                            codeset_name = c('dx_hypertension', 'visit_edip'),
                            filter_logic = c(NA, NA))

#' Execute `prc_process` function
#' This example will use the single site, exploratory, cross sectional
#' configuration
prc_process_example <- prc_process(cohort = cohort,
                                   multi_or_single_site = 'single',
                                   anomaly_or_exploratory = 'exploratory',
                                   time = FALSE,
                                   omop_or_pcornet = 'omop',
                                   prc_event_file = prc_events)

prc_process_example

#' Execute `prc_output` function
prc_output_example <- prc_output(process_output = prc_process_example,
                                 output_function = 'prc_ss_exp_cs')

prc_output_example

#' Easily convert the graph into an interactive ggiraph or plotly object with
#' `make_interactive_ssdqa()`

make_interactive_ssdqa(prc_output_example)
}
